Refactor(policy): 사용자 데이터 제한량 수정 요청 DTO 필드 타입 수정

+++ src/main/java/com/pooli/policy/domain/dto/response/LimitPolicyResDto.java
@@ -16,5 +16,5 @@ public class LimitPolicyResDto {
     private Long lineId;
 
     @Schema(description = "정책 값", example = "1024")
-    private Integer policyValue;
+    private Long policyValue;
 }
